+++ jkube-kit/resource/helm/src/test/java/org/eclipse/jkube/kit/resource/helm/HelmServiceTest.java
@@ -17,19 +17,20 @@ import java.io.File;
 import java.io.IOException;
 
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Properties;
 
+import com.fasterxml.jackson.core.type.TypeReference;
+import org.assertj.core.api.InstanceOfAssertFactories;
 import org.eclipse.jkube.kit.common.JKubeConfiguration;
 import org.eclipse.jkube.kit.common.JavaProject;
 import org.eclipse.jkube.kit.common.KitLogger;
 import org.eclipse.jkube.kit.common.Maintainer;
 import org.eclipse.jkube.kit.common.RegistryConfig;
-import org.eclipse.jkube.kit.common.ResourceFileType;
-import org.eclipse.jkube.kit.common.util.ResourceUtil;
 import org.eclipse.jkube.kit.common.util.Serialization;
 import org.eclipse.jkube.kit.config.resource.ResourceServiceConfig;
 import org.eclipse.jkube.kit.resource.helm.HelmConfig.HelmType;
@@ -37,17 +38,11 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
-import org.mockito.ArgumentCaptor;
-import org.mockito.MockedStatic;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
+import static org.assertj.core.api.Assertions.entry;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.ArgumentMatchers.notNull;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.mockStatic;
-import static org.mockito.Mockito.when;
 
 class HelmServiceTest {
 
@@ -55,10 +50,20 @@ class HelmServiceTest {
   private JKubeConfiguration jKubeConfiguration;
   private ResourceServiceConfig resourceServiceConfig;
   private HelmService helmService;
+  private Path helmSourceDirectory;
+  private Path helmOutputDirectory;
 
   @BeforeEach
-  void setUp() {
-    helmConfig = HelmConfig.builder();
+  void setUp(@TempDir Path tempDir) throws IOException {
+    helmSourceDirectory = Files.createDirectory(tempDir.resolve("helm-source"));
+    Files.createDirectory(helmSourceDirectory.resolve("kubernetes"));
+    Files.createFile(helmSourceDirectory.resolve("kubernetes").resolve("deployment.yaml"));
+    helmOutputDirectory = Files.createDirectory(tempDir.resolve("helm-output"));
+    Files.createDirectory(helmOutputDirectory.resolve("kubernetes"));
+    helmConfig = HelmConfig.builder()
+      .sourceDir(helmSourceDirectory.toFile().getAbsolutePath())
+      .outputDir(helmOutputDirectory.toFile().getAbsolutePath())
+      .tarballOutputDir(helmOutputDirectory.toFile().getAbsolutePath());
     jKubeConfiguration = JKubeConfiguration.builder()
       .project(JavaProject.builder().properties(new Properties()).build())
       .registryConfig(RegistryConfig.builder().settings(new ArrayList<>()).build()).build();
@@ -73,10 +78,9 @@ class HelmServiceTest {
   }
 
   @Test
-  void prepareSourceDirValid_withNonExistentDirectory_shouldThrowException() {
-    File file = mock(File.class);
+  void prepareSourceDirValid_withNonExistentDirectory_shouldThrowException() throws IOException {
     // Given
-    when(file.isDirectory()).thenReturn(true);
+    Files.createDirectory(helmSourceDirectory.resolve("openshift"));
     // When
     IOException result = assertThrows(IOException.class,
             () -> HelmService.prepareSourceDir(helmConfig.build(), HelmType.OPENSHIFT));
@@ -87,43 +91,26 @@ class HelmServiceTest {
   }
 
   @Test
-  void createChartYaml() throws Exception {
-    try (MockedStatic<ResourceUtil> resourceUtilMockedStatic = mockStatic(ResourceUtil.class)) {
-      File outputDir = Files.createTempDirectory("chart-output").toFile();
-      // Given
-      helmConfig.chart("Chart Name").version("1337");
-      // When
-      helmService.createChartYaml(helmConfig.build(), outputDir);
-      // Then
-      ArgumentCaptor<Chart> argumentCaptor = ArgumentCaptor.forClass(Chart.class);
-      resourceUtilMockedStatic.verify(() -> ResourceUtil.save(notNull(), argumentCaptor.capture(), eq(ResourceFileType.yaml)));
-      assertThat(argumentCaptor.getValue())
-          .hasFieldOrPropertyWithValue("apiVersion", "v1")
-          .hasFieldOrPropertyWithValue("name", "Chart Name")
-          .hasFieldOrPropertyWithValue("version", "1337");
-    }
+  void createChartYaml() throws IOException {
+    // Given
+    helmConfig.types(Collections.singletonList(HelmType.KUBERNETES));
+    helmConfig.apiVersion("v1337").chart("Chart Name").version("1337");
+    // When
+    helmService.generateHelmCharts(helmConfig.build());
+    // Then
+    final Map<String, Object> chartYaml = Serialization.unmarshal(
+      helmOutputDirectory.resolve("kubernetes").resolve("Chart.yaml").toFile(),
+      new TypeReference<Map<String, Object>>() {});
+    assertThat(chartYaml)
+      .contains(
+        entry("apiVersion", "v1337"),
+        entry("name", "Chart Name"),
+        entry("version", "1337")
+      );
   }
 
   @Test
-  void createChartYaml_whenApiVersionProvied() throws Exception {
-    try (MockedStatic<ResourceUtil> resourceUtilMockedStatic = mockStatic(ResourceUtil.class)) {
-      File outputDir = Files.createTempDirectory("chart-output-api-version").toFile();
-      // Given
-      helmConfig.chart("Chart Name").version("1337").apiVersion("v2");
-      // When
-      helmService.createChartYaml(helmConfig.build(), outputDir);
-      // Then
-      ArgumentCaptor<Chart> argumentCaptor = ArgumentCaptor.forClass(Chart.class);
-      resourceUtilMockedStatic.verify(() -> ResourceUtil.save(notNull(), argumentCaptor.capture(), eq(ResourceFileType.yaml)));
-      assertThat(argumentCaptor.getValue())
-              .hasFieldOrPropertyWithValue("apiVersion", "v2")
-              .hasFieldOrPropertyWithValue("name", "Chart Name")
-              .hasFieldOrPropertyWithValue("version", "1337");
-    }
-  }
-
-  @Test
-  void createChartYaml_whenInvalidChartYamlFragmentProvided_thenThrowException(@TempDir File outputDir) throws Exception {
+  void createChartYaml_whenInvalidChartYamlFragmentProvided_thenThrowException(@TempDir File outputDir) {
     // Given
     File fragmentsDir = new File(Objects.requireNonNull(getClass().getResource("/invalid-helm-fragments")).getFile());
     jKubeConfiguration = jKubeConfiguration.toBuilder()
@@ -150,6 +137,7 @@ class HelmServiceTest {
     resourceServiceConfig = ResourceServiceConfig.builder()
       .resourceDirs(Collections.singletonList(fragmentsDir)).build();
     helmConfig
+      .apiVersion("v1")
       .chart("Chart Name")
       .version("1337")
       .description("Description from helmconfig")
@@ -181,28 +169,34 @@ class HelmServiceTest {
 
   @Test
   void createChartYamlWithDependencies() throws Exception {
-    try (MockedStatic<ResourceUtil> resourceUtilMockedStatic = mockStatic(ResourceUtil.class)) {
-      // Given
-      File outputDir = Files.createTempDirectory("chart-outputdir").toFile();
-      HelmDependency helmDependency = new HelmDependency()
-          .toBuilder()
-          .name("nginx")
-          .version("1.2.3.")
-          .repository("repository")
-          .build();
-      helmConfig.chart("Chart Name").version("1337")
-          .dependencies(Collections.singletonList(helmDependency));
-      // When
-      helmService.createChartYaml(helmConfig.build(), outputDir);
-      // Then
-      ArgumentCaptor<Chart> argumentCaptor = ArgumentCaptor.forClass(Chart.class);
-      resourceUtilMockedStatic.verify(() -> ResourceUtil.save(notNull(), argumentCaptor.capture(), eq(ResourceFileType.yaml)));
-      assertThat(argumentCaptor.getValue())
-          .hasFieldOrPropertyWithValue("apiVersion", "v1")
-          .hasFieldOrPropertyWithValue("name", "Chart Name")
-          .hasFieldOrPropertyWithValue("version", "1337")
-          .hasFieldOrPropertyWithValue("dependencies",
-              Collections.singletonList(helmDependency));
+    // Given
+    helmConfig.types(Collections.singletonList(HelmType.KUBERNETES));
+    final HelmDependency helmDependency = new HelmDependency()
+        .toBuilder()
+        .name("nginx")
+        .version("1.2.3.")
+        .repository("repository")
+        .build();
+    helmConfig.apiVersion("v1").chart("Chart Name").version("1337")
+        .dependencies(Collections.singletonList(helmDependency));
+    // When
+    helmService.generateHelmCharts(helmConfig.build());
+    // Then
+    final Map<String, Object> chartYaml = Serialization.unmarshal(
+      helmOutputDirectory.resolve("kubernetes").resolve("Chart.yaml").toFile(),
+      new TypeReference<Map<String, Object>>() {});
+    assertThat(chartYaml)
+      .contains(
+        entry("apiVersion", "v1"),
+        entry("name", "Chart Name"),
+        entry("version", "1337")
+      )
+      .extracting("dependencies").asList().singleElement()
+      .asInstanceOf(InstanceOfAssertFactories.map(String.class, String.class))
+      .contains(
+        entry("name", "nginx"),
+        entry("version", "1.2.3."),
+        entry("repository", "repository")
+      );
     }
-  }
 }